requirements.txt
dj-database-url==0.4.1
Django==2.0.2
gunicorn==19.6.0
psycopg2-binary==2.7.6
whitenoise==3.3.0
feedparser>=1.0
djangorestframework==3.8.2
beautifulsoup4==4.4.1
requests==2.9.1
fuzzywuzzy==0.17.0
